:lipstick: Integrate Tailwind CSS and update related configurations
Deleted the old CSS file and updated the currently existing CSS file to accommodate for the use of Tailwind CSS. Also, configured prettier, eslint, and postcss to ensure a standardized coding style and format throughout the project. Made several adjustments in package.json and tsconfig.json to align with these changes. Finally, added related documentation in the README.md file.